fix: 내 순위 공동 순위일 때 반영

+++ app/src/main/java/com/rocket/cosmic_detox/data/model/Ranking.kt
@@ -3,7 +3,7 @@ package com.rocket.cosmic_detox.data.model
 data class RankingInfo(
     val name: String = "",
     val point: Int = 0,
-    val totalTime: Int = 0,
+    val totalTime: Long = 0,
     val uid: String = "",
     val rank: Int = 0
 )
